Move CLIFileindexerProgressDialog to correct package
